+++ relay-dynamic-config/src/feature.rs
@@ -106,6 +106,10 @@ pub enum Feature {
     /// Add a random trace ID to events that lack one.
     #[serde(rename = "organizations:relay-default-trace-id")]
     AddDefaultTraceID,
+    /// Enable experimental expansion of the unreal report in the endpoint rather than in the
+    /// processor. Only enable for organizations with sufficient attachment quota.
+    #[serde(rename = "organizations:relay-unreal-endpoint-expansion")]
+    UnrealEndpointExpansion,
 
     /// Enables OTLP spans to use the Span V2 processing pipeline in Relay.
     ///