+++ relay-event-normalization/src/event.rs
@@ -33,7 +33,7 @@ use crate::span::tag_extraction::extract_span_tags_from_event;
 use crate::utils::{self, MAX_DURATION_MOBILE_MS, get_event_user_tag};
 use crate::{
     BorrowedSpanOpDefaults, BreakdownsConfig, CombinedMeasurementsConfig, GeoIpLookup, MaxChars,
-    ModelCosts, PerformanceScoreConfig, RawUserAgentInfo, SpanDescriptionRule,
+    ModelMetadata, PerformanceScoreConfig, RawUserAgentInfo, SpanDescriptionRule,
     TransactionNameConfig, breakdowns, event_error, legacy, mechanism, remove_other, schema, span,
     stacktrace, transactions, trimming, user_agent,
 };
@@ -134,8 +134,8 @@ pub struct NormalizationConfig<'a> {
     /// Configuration for generating performance score measurements for web vitals
     pub performance_score: Option<&'a PerformanceScoreConfig>,
 
-    /// Configuration for calculating the cost of AI model runs
-    pub ai_model_costs: Option<&'a ModelCosts>,
+    /// Metadata for AI models including costs and context size.
+    pub ai_model_metadata: Option<&'a ModelMetadata>,
 
     /// An initialized GeoIP lookup.
     pub geoip_lookup: Option<&'a GeoIpLookup>,
@@ -192,7 +192,7 @@ impl Default for NormalizationConfig<'_> {
             span_description_rules: Default::default(),
             performance_score: Default::default(),
             geoip_lookup: Default::default(),
-            ai_model_costs: Default::default(),
+            ai_model_metadata: Default::default(),
             enable_trimming: false,
             measurements: None,
             normalize_spans: true,
@@ -321,7 +321,7 @@ fn normalize(event: &mut Event, meta: &mut Meta, config: &NormalizationConfig) {
             .get_or_default::<PerformanceScoreContext>()
             .score_profile_version = Annotated::new(version);
     }
-    enrich_ai_event_data(event, config.ai_model_costs);
+    enrich_ai_event_data(event, config.ai_model_metadata);
     normalize_breakdowns(event, config.breakdowns_config); // Breakdowns are part of the metric extraction too
     normalize_default_attributes(event, meta, config);
     normalize_trace_context_tags(event);
@@ -1540,7 +1540,7 @@ mod tests {
     use serde_json::json;
 
     use super::*;
-    use crate::{ClientHints, MeasurementsConfig, ModelCostV2};
+    use crate::{ClientHints, MeasurementsConfig, ModelCostV2, ModelMetadataEntry};
 
     const IOS_MOBILE_EVENT: &str = r#"
         {
@@ -2306,27 +2306,33 @@ mod tests {
         normalize_event(
             &mut event,
             &NormalizationConfig {
-                ai_model_costs: Some(&ModelCosts {
-                    version: 2,
+                ai_model_metadata: Some(&ModelMetadata {
+                    version: 1,
                     models: HashMap::from([
                         (
                             Pattern::new("claude-2.1").unwrap(),
-                            ModelCostV2 {
-                                input_per_token: 0.01,
-                                output_per_token: 0.02,
-                                output_reasoning_per_token: 0.03,
-                                input_cached_per_token: 0.0,
-                                input_cache_write_per_token: 0.0,
+                            ModelMetadataEntry {
+                                costs: Some(ModelCostV2 {
+                                    input_per_token: 0.01,
+                                    output_per_token: 0.02,
+                                    output_reasoning_per_token: 0.03,
+                                    input_cached_per_token: 0.0,
+                                    input_cache_write_per_token: 0.0,
+                                }),
+                                context_size: None,
                             },
                         ),
                         (
                             Pattern::new("gpt4-21-04").unwrap(),
-                            ModelCostV2 {
-                                input_per_token: 0.02,
-                                output_per_token: 0.03,
-                                output_reasoning_per_token: 0.04,
-                                input_cached_per_token: 0.0,
-                                input_cache_write_per_token: 0.0,
+                            ModelMetadataEntry {
+                                costs: Some(ModelCostV2 {
+                                    input_per_token: 0.02,
+                                    output_per_token: 0.03,
+                                    output_reasoning_per_token: 0.04,
+                                    input_cached_per_token: 0.0,
+                                    input_cache_write_per_token: 0.0,
+                                }),
+                                context_size: None,
                             },
                         ),
                     ]),
@@ -2425,27 +2431,33 @@ mod tests {
         normalize_event(
             &mut event,
             &NormalizationConfig {
-                ai_model_costs: Some(&ModelCosts {
-                    version: 2,
+                ai_model_metadata: Some(&ModelMetadata {
+                    version: 1,
                     models: HashMap::from([
                         (
                             Pattern::new("claude-2.1").unwrap(),
-                            ModelCostV2 {
-                                input_per_token: 0.01,
-                                output_per_token: 0.02,
-                                output_reasoning_per_token: 0.03,
-                                input_cached_per_token: 0.04,
-                                input_cache_write_per_token: 0.0,
+                            ModelMetadataEntry {
+                                costs: Some(ModelCostV2 {
+                                    input_per_token: 0.01,
+                                    output_per_token: 0.02,
+                                    output_reasoning_per_token: 0.03,
+                                    input_cached_per_token: 0.04,
+                                    input_cache_write_per_token: 0.0,
+                                }),
+                                context_size: None,
                             },
                         ),
                         (
                             Pattern::new("gpt4-21-04").unwrap(),
-                            ModelCostV2 {
-                                input_per_token: 0.09,
-                                output_per_token: 0.05,
-                                output_reasoning_per_token: 0.0,
-                                input_cached_per_token: 0.0,
-                                input_cache_write_per_token: 0.0,
+                            ModelMetadataEntry {
+                                costs: Some(ModelCostV2 {
+                                    input_per_token: 0.09,
+                                    output_per_token: 0.05,
+                                    output_reasoning_per_token: 0.0,
+                                    input_cached_per_token: 0.0,
+                                    input_cache_write_per_token: 0.0,
+                                }),
+                                context_size: None,
                             },
                         ),
                     ]),
@@ -2527,16 +2539,19 @@ mod tests {
         normalize_event(
             &mut event,
             &NormalizationConfig {
-                ai_model_costs: Some(&ModelCosts {
-                    version: 2,
+                ai_model_metadata: Some(&ModelMetadata {
+                    version: 1,
                     models: HashMap::from([(
                         Pattern::new("claude-2.1").unwrap(),
-                        ModelCostV2 {
-                            input_per_token: 0.01,
-                            output_per_token: 0.02,
-                            output_reasoning_per_token: 0.03,
-                            input_cached_per_token: 0.0,
-                            input_cache_write_per_token: 0.0,
+                        ModelMetadataEntry {
+                            costs: Some(ModelCostV2 {
+                                input_per_token: 0.01,
+                                output_per_token: 0.02,
+                                output_reasoning_per_token: 0.03,
+                                input_cached_per_token: 0.0,
+                                input_cache_write_per_token: 0.0,
+                            }),
+                            context_size: None,
                         },
                     )]),
                 }),
@@ -2599,27 +2614,33 @@ mod tests {
         normalize_event(
             &mut event,
             &NormalizationConfig {
-                ai_model_costs: Some(&ModelCosts {
-                    version: 2,
+                ai_model_metadata: Some(&ModelMetadata {
+                    version: 1,
                     models: HashMap::from([
                         (
                             Pattern::new("claude-2.1").unwrap(),
-                            ModelCostV2 {
-                                input_per_token: 0.01,
-                                output_per_token: 0.02,
-                                output_reasoning_per_token: 0.0,
-                                input_cached_per_token: 0.04,
-                                input_cache_write_per_token: 0.0,
+                            ModelMetadataEntry {
+                                costs: Some(ModelCostV2 {
+                                    input_per_token: 0.01,
+                                    output_per_token: 0.02,
+                                    output_reasoning_per_token: 0.0,
+                                    input_cached_per_token: 0.04,
+                                    input_cache_write_per_token: 0.0,
+                                }),
+                                context_size: None,
                             },
                         ),
                         (
                             Pattern::new("gpt4-21-04").unwrap(),
-                            ModelCostV2 {
-                                input_per_token: 0.09,
-                                output_per_token: 0.05,
-                                output_reasoning_per_token: 0.06,
-                                input_cached_per_token: 0.0,
-                                input_cache_write_per_token: 0.0,
+                            ModelMetadataEntry {
+                                costs: Some(ModelCostV2 {
+                                    input_per_token: 0.09,
+                                    output_per_token: 0.05,
+                                    output_reasoning_per_token: 0.06,
+                                    input_cached_per_token: 0.0,
+                                    input_cache_write_per_token: 0.0,
+                                }),
+                                context_size: None,
                             },
                         ),
                     ]),
@@ -2686,8 +2707,8 @@ mod tests {
         normalize_event(
             &mut event,
             &NormalizationConfig {
-                ai_model_costs: Some(&ModelCosts {
-                    version: 2,
+                ai_model_metadata: Some(&ModelMetadata {
+                    version: 1,
                     models: HashMap::new(),
                 }),
                 ..NormalizationConfig::default()
@@ -2730,8 +2751,8 @@ mod tests {
         normalize_event(
             &mut event,
             &NormalizationConfig {
-                ai_model_costs: Some(&ModelCosts {
-                    version: 2,
+                ai_model_metadata: Some(&ModelMetadata {
+                    version: 1,
                     models: HashMap::new(),
                 }),
                 ..NormalizationConfig::default()